refactor(ui_archi): regroup visual .vue into ui folder
