--- yaki_admin/src/features/customer/CustomerPage.vue
@@ -1,17 +1,0 @@
-<script setup lang="ts">
-import PageLayout from "@/layouts/PageGeneralLayout.vue";
-import SideBarCustomer from "@/features/customer/components/SideBarCustomer.vue";
-import {RouterView} from "vue-router";
-</script>
-
-<template>
-  <page-layout>
-    <template v-slot:sidebarContent>
-      <side-bar-customer />
-    </template>
-
-    <template v-slot:pageContent>
-      <router-view />
-    </template>
-  </page-layout>
-</template>
