+++ yaki_admin/src/features/customer/layouts/CustomerPageContentTeamList.vue
@@ -6,9 +6,8 @@ import { useTeamStore } from "@/stores/teamStore";
 import modalState from "@/features/modal/services/modalState";
 import { MODALMODE } from "@/constants/modalMode";
 
-import PageContentLayout from "@/layouts/PageContentLayout.vue";
+import PageContentLayout from "@/ui/layouts/PageContentLayout.vue";
 import SideBarButton from "@/features/shared/components/SideBarButton.vue";
-import TeamWithCaptainCard from "@/features/customer/components/TeamWithCaptainCard.vue";
 import plusIcon from "@/assets/images/plus.png";
 import { useCaptainStore } from "@/stores/captainStore";
 import { useSelectedRoleStore } from "@/stores/selectedRole";
