+++ yaki_admin/src/ui/components/LoggedUserCard.vue
