+++ yaki_admin/src/ui/layouts/PageContentLayout.vue
