+++ yaki_admin/src/ui/views/captain/CaptainNoTeamView.vue
